requirements.txt
aws-lambda-powertools==2.38.1
python-dotenv==1.0.1
requests==2.32.3
twilio==9.2.3
pydantic==2.8.2
openai==1.51.0
boto3==1.34.154
httpx<0.28
